backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
boto3==1.28.64
python-multipart==0.0.6
sqlalchemy<2.0.0,>=1.4.0
sqlalchemy-redshift==0.8.14
pandas==2.1.1
